exit properly and raspberry led changed
Before the exit_clean was not exiting at all.
Now clean_exit will throw exit exception.

The Raspberry LED was previously linked to the 3,3 PWR.
This PWR is always on, even after shutdown.
As this is useless, now the behavior is to have a dedicated
GPIO that will be turned on while the bell-controller is running.